requirements.txt
requests==2.28.2
pandas==1.5.3
numpy==1.24.2
python-binance==1.0.17
openpyxl==3.0.7